refactor: update import path and virtualization state names
